Immediately go to exact match on tags
It seems like this is probably an option somewhere in etags-select.el,
or at least in Emacs's own etags.el, but I couldn't find it, so I
wrote it.

Could probably use a customize variable to let people pick whether or
not "jump immediately without prompting" is the default behavior.