Fix completion not working sometimes
I don't really have an explanation for *when* completion didn't work,
but it definitely didn't work sometimes, and I'm pretty sure that's
because it was trying to read the tags-completion-table without the
tags buffer being the current buffer.  etags.el (with Emacs 25.2, at
least) sets up tags-completion-table as a buffer-local variable in the
tags file's buffer itself, and I think it probably only updates that
buffer-local variable as well, so calling tags-completion-table
outside the tags buffer is a no-no.